Add table numbers to document pdf
